+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 * [HexaPDF::Type::OutlineItem#open?] for getting the open state of an outline
   item
 
+### Fixed
+
+* Decryption of invalid files having empty strings or streams when using the AES
+  algorithm
+
 
 ## 0.31.0 - 2023-02-22